adds nodes folder
