--- libraries/Pliant/IAndNode.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Pliant
-{
-    public interface IAndNode
-    {
-        IReadOnlyList<INode> Children { get; } 
-    }
-}
