--- libraries/Pliant/INode.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pliant
-{
-    public interface INode
-    {
-        int Origin { get; }
-
-        int Location { get; }
-
-        NodeType NodeType { get; }
-    }
-}
