+++ libraries/Pliant/IState.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Pliant.Nodes;
+using System;
 namespace Pliant
 {
     public interface IState
