--- libraries/Pliant/NodeType.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pliant
-{
-    public enum NodeType
-    {
-        Intermediate,
-        Symbol,
-        Terminal,
-        Token,
-        Virtual
-    }
-}
