+++ libraries/Pliant/Nodes/IAndNode.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Pliant.Nodes
+{
+    public interface IAndNode
+    {
+        IReadOnlyList<INode> Children { get; } 
+    }
+}
