+++ libraries/Pliant/Nodes/IIntermediateNode.cs
@@ -1,0 +1,7 @@
+﻿namespace Pliant.Nodes
+{
+    public interface IIntermediateNode : IInternalNode
+    {
+        IState State { get; }
+    }
+}
