+++ libraries/Pliant/Nodes/INode.cs
@@ -1,0 +1,11 @@
+﻿namespace Pliant.Nodes
+{
+    public interface INode
+    {
+        int Origin { get; }
+
+        int Location { get; }
+
+        NodeType NodeType { get; }
+    }
+}
