+++ libraries/Pliant/Nodes/ISymbolNode.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Pliant.Nodes
+{
+    public interface ISymbolNode : IInternalNode
+    {
+        ISymbol Symbol { get;}
+    }
+}
