+++ libraries/Pliant/Nodes/ITokenNode.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Pliant.Nodes
+{
+    public interface ITokenNode : INode
+    {
+        IToken Token { get; }
+    }
+}
