+++ libraries/Pliant/Nodes/NodeType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Pliant.Nodes
+{
+    public enum NodeType
+    {
+        Intermediate,
+        Symbol,
+        Terminal,
+        Token,
+        Virtual
+    }
+}
