+++ libraries/Pliant/ParseEngine.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Pliant.Nodes;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
