+++ libraries/Pliant/PulseRecognizer.cs
@@ -1,4 +1,5 @@
-﻿using Pliant.Terminals;
+﻿using Pliant.Nodes;
+using Pliant.Terminals;
 using System.Diagnostics;
 using System.Linq;
