+++ libraries/Pliant/RegexParser.cs
@@ -1,4 +1,5 @@
-﻿using Pliant.Terminals;
+﻿using Pliant.Nodes;
+using Pliant.Terminals;
 using System;
 using System.Collections.Generic;
 using System.Linq;
